requirements.txt
discord.py
motor
dnspython
python-dotenv
pymongo